requirements.txt
flask
python-dotenv
pydub
torch
transformers
accelerate
ffmpeg-python
